+++ docs/api/changelog.rst
@@ -3,6 +3,13 @@
 Changelog
 =========
 
+.. _release-2.2.0:
+
+2.2.0 - TBD
+   * Add a mypy plugin
+   * The extra ``__testname__`` and ``is_noy_spec`` lines now get a type:ignore
+     comment so that mypy doesn't complain about them
+
 .. _release-2.1.0:
 
 2.1.0 - 6 November 2021